+++ collector-media3-exoplayer/src/main/java/com/bitmovin/analytics/media3/exoplayer/Media3ExoPlayerAdapter.kt
@@ -126,10 +126,6 @@ internal class Media3ExoPlayerAdapter(
             return exoplayerContext.position
         }
 
-    override fun clearValuesAfterSendingOfSample() {
-        meter.reset()
-    }
-
     private fun startup(position: Long) {
         qualityEventDataManipulator.setFormatsFromPlayerOnStartup()
         stateMachine.transitionState(PlayerStates.STARTUP, position)