+++ collector/src/main/java/com/bitmovin/analytics/adapters/PlayerAdapter.kt
@@ -22,10 +22,6 @@ interface PlayerAdapter {
 
     fun resetSourceRelatedState()
 
-    // Called after sending EventData
-    // Should clear all values that are only needed per sample
-    fun clearValuesAfterSendingOfSample()
-
     fun createEventData(): EventData
 
     fun createEventDataForCustomDataEvent(sourceMetadata: SourceMetadata): EventData