+++ collector/src/test/java/com/bitmovin/analytics/utils/DownloadSpeedMeterTest.kt
@@ -1,46 +1,48 @@
 package com.bitmovin.analytics.utils
 
-import com.bitmovin.analytics.data.SpeedMeasurement
 import org.assertj.core.api.Assertions
+import org.junit.Before
 import org.junit.Test
 
 class DownloadSpeedMeterTest {
-    var measurement1 = SpeedMeasurement(20, 1000)
-    var measurement2 = SpeedMeasurement(30, 1500)
-    var measurement3 = SpeedMeasurement(50, 2000)
-    var measurement4 = SpeedMeasurement(20, 2000)
-    var measurement5 = SpeedMeasurement(20, 1500)
-    var meter = DownloadSpeedMeter()
+    private var measurement1 = DownloadSpeedMeasurement(20, 1000)
+    private var measurement2 = DownloadSpeedMeasurement(30, 1500)
+    private var measurement3 = DownloadSpeedMeasurement(50, 2000)
+    private var measurement4 = DownloadSpeedMeasurement(20, 2000)
+    private var measurement5 = DownloadSpeedMeasurement(20, 1500)
+    private var meter = DownloadSpeedMeter()
+
+    @Before
+    fun setup() {
+        meter.reset()
+    }
 
     @Test
     fun testMeterAddAndRest() {
-        meter.reset()
         meter.addMeasurement(measurement1)
         meter.addMeasurement(measurement2)
         meter.addMeasurement(measurement3)
         meter.addMeasurement(measurement4)
         meter.addMeasurement(measurement5)
 
-        var info = meter.getInfo()
-        // total of 5 measurements
+        var info = meter.getInfoAndReset()
+
         Assertions.assertThat(info.segmentsDownloadCount).isEqualTo(5)
 
-        meter.reset()
-        info = meter.getInfo()
-        // total of 5 measurements
+        info = meter.getInfoAndReset()
+
         Assertions.assertThat(info.segmentsDownloadCount).isEqualTo(0)
     }
 
     @Test
     fun testSpeedMeasurements() {
-        meter.reset()
         meter.addMeasurement(measurement1)
         meter.addMeasurement(measurement2)
         meter.addMeasurement(measurement3)
         meter.addMeasurement(measurement4)
         meter.addMeasurement(measurement5)
 
-        val info = meter.getInfo()
+        val info = meter.getInfoAndReset()
         // total of 5 measurements
         Assertions.assertThat(info.segmentsDownloadCount).isEqualTo(5)
         // sum of all durations
@@ -48,9 +50,24 @@ class DownloadSpeedMeterTest {
         // sum of all sizes
         Assertions.assertThat(info.segmentsDownloadSize).isEqualTo(8000)
         // slowest download -> measurement4
-        Assertions.assertThat(info.minDownloadSpeed).isEqualTo(800.0f)
+        Assertions.assertThat(info.minDownloadSpeed).isEqualTo(320.0f)
         // fastest download -> measurement3
-        Assertions.assertThat(info.maxDownloadSpeed).isEqualTo(320.0f)
+        Assertions.assertThat(info.maxDownloadSpeed).isEqualTo(800.0f)
         Assertions.assertThat(info.avgDownloadSpeed).isEqualTo(504.0f)
+    }
+
+    @Test
+    fun `reset clears the value after calling`() {
+        meter.addMeasurement(measurement1)
+        meter.addMeasurement(measurement2)
+        meter.addMeasurement(measurement3)
+        meter.addMeasurement(measurement4)
+        meter.addMeasurement(measurement5)
+
+        meter.reset()
+
+        val info = meter.getInfoAndReset()
+
+        Assertions.assertThat(info.segmentsDownloadCount).isEqualTo(0)
     }
 }